Changed command from string to symbol, for performance improvements.
In the memory test, added a call to display the contents of the doc, which takes a fraction of a second against this code.  Running the same call against the example code that used blocks (ali_rizvi_fixed) takes many seconds (approx 30 seconds on my machine.)  In order to pick which approach is best, ask the consumer the following questions:
How often do you need to manipulate the data?  How often do you need to view it?
What are your performance expectations for viewing the data? What are the expectations for manipulating the data?